Misc perf improvements

+++ src/filter.rs
@@ -1,3 +1,3 @@
 pub trait Filter {
-    fn process(&mut self, in_sample: f32, sample_rate: f32, cutoff: f64, resonance: f64) -> f32;
+    fn process(&mut self, in_sample: f32, sample_rate: f32, cutoff: f32, resonance: f32) -> f32;
 }
